tests/wpt/mozilla/tests/mozilla/has-instance.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var obj = {"__proto__": Element.prototype};
  assert_true(obj instanceof Node);
}, "Manually-constructed prototype chains are correctly handled by instanceof");
</script>
